allow change consumers to mark progress at any time
Now, change consumers receive a ProgressReporter when they are created,
and can use it to save progress at any point of their lifetime. This was
made to give flexibility to change consumers in terms of when they want
to save progress for a stream, even outside Consume() calls.

Added an utility class - PeriodicProgressReporter, which wraps a
ProgressReporter and reports progress in fixed intervals.