topology.go: use at most CL=QUORUM for querying generations
Consistency level for querying generation tables was chosen as follows:

- cluster size == 1 => CL=ONE,
- cluster size == 2 => CL=QUORUM,
- cluster size >= 3 => CL=ALL.

Replication factor for the generations table is 3, so choosing CL=ONE
and CL=QUORUM for tiny clusters allows the library to work with them.
However, choosing CL=ALL for other sizes is an overkill - the library
won't tolerate any nodes being down. Moreover, the generation tables
themselves are written to with CL=QUORUM, so we should get similar
results when using CL=QUORUM for bigger cluster sizes as well.

This commit changes the algorithm so that CL=QUORUM is used for clusters
of size 3 and larger.