introduce internal/testutils package
As a preparation for setting up CI, an internal package with test
utilities is added. The new utilities make it easier to run the tests
automatically: they unify how the tests connect to the
source/destination cluster and allow generating unique names/keyspaces
so that tests that run in parallel don't interfere with each other.

The new utilities allow for the following:

- generating a unique name,
- getting information about the contact point of the source/destination
  cluster,
- creating a keyspace (optionally with a unique name).

The existing tests are modified to use the new utilities.