change: add ChangeOrEmptyNotificationConsumer extension interface
Currently, the responsibility for saving progress is put on the
user-provided ChangeConsumer. The idea was that it would let the user
control when the progress would actually be saved. However, the
ChangeConsumer is only notified about the the progress when more rows
appear in the stream it is associated with. If there is a long period of
time when no rows appear on the stream, the ChangeConsumer won't be able
to save information about this progress. If the application is restarted
during such a long period of time, the ChangeConsumer for that stream
will have to start from the beginning of the period.

This commit offers a way out by adding an extension interface:
ChangeOrEmptyNotificationConsumer. It adds an Empty() method which gets
called when the library has read an empty window of rows, and provides
the consumer with information about the window's end. The implementors
of this method can report progress there - it's not an automatic fix,
unfortunately.

A test is added which verifies that the issue is fixed.

Co-authored-by: hartmut <info@hartmut.co.uk>