progress: prevent TBPM from restarting reading stream from generation start
The ProgressManager interface is used as follows:

1. GetCurrentGeneration is called by the library to fetch information
   about the current generation,
2. The library chooses the generation to read from and calls
   StartGeneration,
3. For each stream, GetProgress is called to determine from which
   timestamp to start reading,
4. The user's change consumers are supposed to periodically call
   SaveProgress to mark progress for the stream,
5. When the generation ends, library waits for all streams to be read to
   end, then choose the next generation and goto 2.

There is a problem with the TableBackedProgressManager if we reach 3 but
do not execute 4. When the application is restarted, the TBPM will
correctly read the saved generation, but it will not find saved progress
for streams and the GetProgress invocations will return a zero
timestamp. According to the contract of the GetProgress function, the
library will start reading the stream starting from the beginning of the
generation - and this is rarely something what we would like to do, for
example if the first run of the application was done on a very old
cluster.

This commit fixes the issue by extending the ProgressManager interface.
The new extension is named ProgressManagerWithStartTime and allows to
store/retrieve the timestamp from which the application originally
started reading. The log reader will not attempt to read data earlier
than the saved start timestamp.

The new interface is implemented for the TableBackedProgressManager. The
commit also adds a test which verifies that the issue is fixed.