reader: adjust reading start time for the TTL
The CDC tables usually have a "default TTL" option which controls the
TTL of the rows that are inserted to it. It does not make much sense to
try to read changes that are older than TTL because they are
automatically expired by the database.

This commit adjusts the reading start time based on the tables' TTLs.