move splitStreams to generationSource
move the splitStreams operation from the reader to generationSource,
because it depends on whether the generation source is for tablets or
vnodes and we want to have different implementations for it.

until now, the streams are always grouped by the vnode index encoded in
the stream id. however, for tablets, the vnode index is meaningless and
it's always set to 0, so it will put all streams in a single batch. this
is undesired.

we want to keep this implementation for vnode-based tables, but have a
different implementation for tablets. for this, we move the splitStreams
operation to be done by the generationSource, and change it to return
the grouped streams, instead of the reader doing it.

for the tablets generation source, we return each stream in a separate
group.